Move exception from message to text attachment
Sometimes Chef will have a lengthy exception stack trace.  When this happens on 2 or 3 hosts it can flood a slack channel with a lot of text.  Moving the exception from the message body to a text attachment allows slack to collapse the message and show a "show more" link.

See the `text` section of https://api.slack.com/docs/attachments for more info