Fix reporting exceptions with libyajl2
Some chef versions, like 11.14.2, may end up installing libyajl2, instead of
yajl-ruby, and it seems they treat exceptions differently. The end result is
that the webhook fails with "invalid_payload" and the json seems to be invalid.

By just doing a "to_s", it always work (it works fine on a string too).